remove publications view from dashboard, move dataset fetching to dataset_view
